feat(game): rebuild the trial, jury and prison flow on the shipped crime data
A crime could not be tried: the arrest that starts the flow was never raised, the jury was
never given a case to read, and the verdict window never opened for a juror because the
client was never told it was one. This rebuilds the path against what the client's own
court UI expects.

Arrest and sentence (JusticeManager, ArrestRules)

* The arrest-state buff the shipped skills put on a criminal (4906 / 4962) starts the
  escort; it runs for the buff's own length and lands them at their faction's courthouse,
  where the imprison-or-trial offer goes out. A wanted character who dies comes back at
  court instead, as retail treats that death as the summons.
* A sentence is the shipped prisoner buff (631, 30 minutes): its own timeout returns the
  prisoner to the prison exit, so nothing has to release them by hand. Serving it pays the
  crime down by the shipped reduction so the wanted state ends with it, and a prisoner is
  locked out of instanced content while they serve.

Trial (TrialManager, TrialVerdictRules, TrialSeatRules, TrialTimingRules, TrialSentenceRules,
TrialAudienceRules)

* Phases are the client's own trial states (0 free .. 8 post-sentence) and every phase goes
  out with the seconds it has left, so the client's clocks run down with ours; a phase
  token keeps a stale timer from moving a later phase.
* The bench: a juror answers the invite, is promised a chair, is summoned to it and is
  seated by the client's own answer to that summon - the summon is also what makes the
  client a juror and arms its verdict window. Seating teleports them onto the chair the
  court names and sits them in the court's pose.
* The verdict is the row the client sends: row 1 is not guilty and rows 2-6 are the guilty
  tiers, so the tally has to read the row and not the sentence. The ruling goes out once
  every seat has voted (or the bench is empty, where the defendant's own record stands),
  and the trial is closed the way the client expects: ruling status, post-sentence, ruling
  closed.
* A courthouse hears one case at a time, so a second defendant queues with their place in
  the queue shown in their wait window, and the next case opens when the courtroom is free.
* An onlooker watches the case the bench does: joining the gallery hands them the same case
  file, tally and phase, without which every tab of their court window is empty (measured:
  only the audience-joined packet used to arrive).
* A defendant or juror who logs out - or crashes - leaves the case cleanly on the same path
  the family and expedition managers use.

Crime records

* A case file is the defendant's crime events, sent as the court's crime records with the
  "no case file yet" total (-1) until the bench convenes, and the crimes a case was served
  for are dropped when it closes.

Also here: the free-attack toggle the client casts once per press (the shipped effect row
carries no direction, so it toggles), the prison mounds' evidence loot, and the stale packet
whose opcode collided with the trial block.

Tests: the verdict/seat/sentence/audience rules, the trial packet shapes and opcodes against
the client's own table, and the crime-record packet layout.

+++ AAEmu.Game/Core/Packets/G2C/SCChangeTrialStatePacket.cs
@@ -3,7 +3,7 @@ using AAEmu.Game.Core.Network.Game;
 
 namespace AAEmu.Game.Core.Packets.G2C;
 
-public class SCChangeTrialStatePacket(uint trialId, byte state, int curJury, uint remainTime)
+public class SCChangeTrialStatePacket(ulong trialId, byte state, int curJury, uint remainTime)
     : GamePacket(SCOffsets.SCChangeTrialStatePacket, 1)
 {
     public override PacketStream Write(PacketStream stream)
