+++ AAEmu.Game/Core/Packets/G2C/SCTrialInfoPacket.cs
@@ -4,7 +4,7 @@ using AAEmu.Game.Core.Network.Game;
 namespace AAEmu.Game.Core.Packets.G2C;
 
 public class SCTrialInfoPacket(
-    uint type,
+    ulong type,
     int crimePoint,
     int arrest,
     int acceptGuilty,
